Frictionfull Space (#514)
# Description
Makes it so that the station and the ATS get a very tiny bit of friction
to prevent cargo tech pros from sending either of those out of this
galaxy cluster (which has actually happened multiple times on two
servers and required either admin intervention or early round ending).

# Technical details
Added a PassiveDampeningComponent which defines how much friction an
entity receives while in 0g. FrictionRemoverSystem was updated to try to
fetch this component from an entity before updating its dampening. A new
system was added to automatically add this component (if it's not
already defined) to all station grids.

# Media
See the #when-you-code-it channel for a preview. It's kinda hard to
demonstrate, but after a few minutes, stations and the ATS come to an
almost complete stop.

# Changelog
:cl:
- tweak: Space stations now have a tiny bit of velocity dampening to
prevent them from being flunged into the void.

(cherry picked from commit e092203d1134521f95c44c424bfa57f1ec7ffe53)